requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Werkzeug==3.1.3
Jinja2==3.1.5
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
pip==23.2.1
setuptools==75.7.0
wheel==0.45.1
cryptography==44.0.0
python-dateutil==2.8.2